+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [v5.3.0](https://github.com/fastly/fastly-ruby/releases/tag/release/v5.3.0) (2023-06-27)
+
+**Enhancements:**
+
+- feat(rate_limiter): implement POST/PUT endpoints.
+
+**Bug fixes:**
+
+- fix(automation_tokens): generate explicit error response type.
+- fix(historical_stats): extract primitive into custom type.
+
 ## [v5.2.2](https://github.com/fastly/fastly-ruby/releases/tag/release/v5.2.2) (2023-06-23)
 
 **Bug fixes:**